// Server/proto/ClientRequest-TCP.ts
export interface LoginRequest {
  username: string;
  password?: Uint8Array | undefined;
}

export interface FindMatchRequest {
  sessionId: Uint8Array;
}

export interface ChatMessageRequest {
  sessionId: Uint8Array;
  chatMessage: string;
}

function createBaseLoginRequest(): LoginRequest {
  return { username: "", password: undefined };
}

export const LoginRequest = {
  encode(message: LoginRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.username !== "") {
      writer.uint32(10).string(message.username);
    }
    if (message.password !== undefined) {
      writer.uint32(18).bytes(message.password);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): LoginRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLoginRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.username = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.password = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LoginRequest {
    return {
      username: isSet(object.username) ? globalThis.String(object.username) : "",
      password: isSet(object.password) ? bytesFromBase64(object.password) : undefined,
    };
  },

  toJSON(message: LoginRequest): unknown {
    const obj: any = {};
    if (message.username !== "") {
      obj.username = message.username;
    }
    if (message.password !== undefined) {
      obj.password = base64FromBytes(message.password);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LoginRequest>, I>>(base?: I): LoginRequest {
    return LoginRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LoginRequest>, I>>(object: I): LoginRequest {
    const message = createBaseLoginRequest();
    message.username = object.username ?? "";
    message.password = object.password ?? undefined;
    return message;
  },
};

function createBaseFindMatchRequest(): FindMatchRequest {
  return { sessionId: new Uint8Array(0) };
}

export const FindMatchRequest = {
  encode(message: FindMatchRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.sessionId.length !== 0) {
      writer.uint32(10).bytes(message.sessionId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): FindMatchRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindMatchRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.sessionId = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindMatchRequest {
    return { sessionId: isSet(object.sessionId) ? bytesFromBase64(object.sessionId) : new Uint8Array(0) };
  },

  toJSON(message: FindMatchRequest): unknown {
    const obj: any = {};
    if (message.sessionId.length !== 0) {
      obj.sessionId = base64FromBytes(message.sessionId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindMatchRequest>, I>>(base?: I): FindMatchRequest {
    return FindMatchRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindMatchRequest>, I>>(object: I): FindMatchRequest {
    const message = createBaseFindMatchRequest();
    message.sessionId = object.sessionId ?? new Uint8Array(0);
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };