// Server/proto/ServerResponse-TCP.ts
export interface LoginResponse {
  success: boolean;
  sessionId?: Uint8Array | undefined;
}

export interface FindMatchResponse {
  success: boolean;
  matchId?: Uint8Array | undefined;
}

function createBaseLoginResponse(): LoginResponse {
  return { success: false, sessionId: undefined };
}

export const LoginResponse = {
  encode(message: LoginResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.success !== false) {
      writer.uint32(8).bool(message.success);
    }
    if (message.sessionId !== undefined) {
      writer.uint32(18).bytes(message.sessionId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): LoginResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLoginResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.success = reader.bool();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.sessionId = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LoginResponse {
    return {
      success: isSet(object.success) ? globalThis.Boolean(object.success) : false,
      sessionId: isSet(object.sessionId) ? bytesFromBase64(object.sessionId) : undefined,
    };
  },

  toJSON(message: LoginResponse): unknown {
    const obj: any = {};
    if (message.success !== false) {
      obj.success = message.success;
    }
    if (message.sessionId !== undefined) {
      obj.sessionId = base64FromBytes(message.sessionId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LoginResponse>, I>>(base?: I): LoginResponse {
    return LoginResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LoginResponse>, I>>(object: I): LoginResponse {
    const message = createBaseLoginResponse();
    message.success = object.success ?? false;
    message.sessionId = object.sessionId ?? undefined;
    return message;
  },
};

function createBaseFindMatchResponse(): FindMatchResponse {
  return { success: false, matchId: undefined };
}

export const FindMatchResponse = {
  encode(message: FindMatchResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.success !== false) {
      writer.uint32(8).bool(message.success);
    }
    if (message.matchId !== undefined) {
      writer.uint32(18).bytes(message.matchId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): FindMatchResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindMatchResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.success = reader.bool();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.matchId = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindMatchResponse {
    return {
      success: isSet(object.success) ? globalThis.Boolean(object.success) : false,
      matchId: isSet(object.matchId) ? bytesFromBase64(object.matchId) : undefined,
    };
  },

  toJSON(message: FindMatchResponse): unknown {
    const obj: any = {};
    if (message.success !== false) {
      obj.success = message.success;
    }
    if (message.matchId !== undefined) {
      obj.matchId = base64FromBytes(message.matchId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindMatchResponse>, I>>(base?: I): FindMatchResponse {
    return FindMatchResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindMatchResponse>, I>>(object: I): FindMatchResponse {
    const message = createBaseFindMatchResponse();
    message.success = object.success ?? false;
    message.matchId = object.matchId ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };